Shopping Mall Management Project

+++ src/main/java/com/example/user/repo/userRepo.java
@@ -3,6 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.user.userdetails;
 
-public interface userRepo  extends JpaRepository<userdetails, Long>{
+public interface userRepo  extends JpaRepository<userdetails, Integer>{
 
 }
